style: Rename 'header-and-jwt' to 'jwt'
