+++ api/modules/auth/controller.ts
@@ -6,8 +6,8 @@ import { generateDefaultTOTP, validateDefaultTOTP } from '../../core/rfc6238';
 import { parseBasicAuth } from '../../core/rfc7617';
 import AppError from '../../util/app-error';
 import getDeviceID from '../../util/device-id';
-import { extractJWT, signJWS, verifyToken } from '../../util/header-and-jwt';
 import isHTTPS from '../../util/is-https';
+import { extractJWT, signJWS, verifyToken } from '../../util/jwt';
 import { verifyPassword } from '../../util/passwords';
 import safeCompare from '../../util/safe-compare';
 import sendResponse from '../../util/send-response';
